DRY the message-text render logic into RenderMessageText.vue
