+++ frontend/views/containers/chatroom/MessageBase.vue
@@ -18,28 +18,13 @@
           span.has-text-1 {{ humanDate(datetime, { hour: 'numeric', minute: 'numeric' }) }}
 
       slot(name='body')
-        p.c-replying.custom-markdown-content(
+        render-message-text.custom-markdown-content(
           v-if='replyingMessage'
+          :isReplyingMessage='true'
+          :text='replyingMessage'
           @click='$emit("reply-message-clicked")'
         )
-          template(v-for='(objReplyMessage, index) in replyMessageObjects')
-            span(
-              v-if='isText(objReplyMessage)'
-              v-safe-html:a='objReplyMessage.text'
-            )
-            span.c-member-mention(
-              v-else-if='isMemberMention(objReplyMessage)'
-              :class='{"c-mention-to-me": objReplyMessage.toMe}'
-            ) {{ objReplyMessage.text }}
-            span.c-channel-mention(
-              v-else-if='isChannelMention(objReplyMessage)'
-              :tabindex='objReplyMessage.disabled ? undefined : 0'
-              :class='{ "is-disabled": objReplyMessage.disabled }'
-              @click='navigateToChatRoom(objReplyMessage)'
-              @keyup.enter='navigateToChatRoom(objReplyMessage)'
-            )
-              i(:class='"icon-" + objText.icon')
-              span {{ objText.text }}
+
         send-area(
           v-if='isEditing'
           :defaultText='swapMentionIDForDisplayname(text)'
@@ -47,28 +32,11 @@
           @send='onMessageEdited'
           @cancelEdit='cancelEdit'
         )
-
-        p.c-text.custom-markdown-content(v-else-if='text')
-          template(v-for='objText in textObjects')
-            span(
-              v-if='isText(objText)'
-              v-safe-html:a='objText.text'
-            )
-            template(v-else-if='isMemberMention(objText)')
-              span.c-mention-profile-card-wrapper(v-if='objText.userID')
-                profile-card(:contractID='objText.userID' direction='top-left')
-                  span.c-member-mention(:class='{"c-mention-to-me": objText.toMe}') {{ objText.text }}
-              span.c-member-mention(v-else class='c-mention-to-me') {{ objText.text }}
-            span.c-channel-mention(
-              v-else-if='isChannelMention(objText)'
-              :tabindex='objText.disabled ? undefined : 0'
-              :class='{ "is-disabled": objText.disabled }'
-              @click='navigateToChatRoom(objText)'
-              @keyup.enter='navigateToChatRoom(objText)'
-            )
-              i(:class='"icon-" + objText.icon')
-              span {{ objText.text }}
-          i18n.c-edited(v-if='edited') (edited)
+        render-message-text.c-text.custom-markdown-content(
+          v-else-if='text'
+          :text='text'
+          :edited='edited'
+        )
 
       .c-attachments-wrapper(v-if='hasAttachments')
         chat-attachment-preview(
@@ -114,33 +82,23 @@
 
 <script>
 import sbp from '@sbp/sbp'
-import { mapGetters } from 'vuex'
 import Avatar from '@components/Avatar.vue'
 import Tooltip from '@components/Tooltip.vue'
 import ProfileCard from '@components/ProfileCard.vue'
 import emoticonsMixins from './EmoticonsMixins.js'
 import MessageActions from './MessageActions.vue'
 import MessageReactions from './MessageReactions.vue'
+import RenderMessageText from './RenderMessageText.vue'
 import SendArea from './SendArea.vue'
 import ChatAttachmentPreview from './file-attachment/ChatAttachmentPreview.vue'
 import { humanDate } from '@model/contracts/shared/time.js'
-import { makeMentionFromUserID, swapMentionIDForDisplayname, makeChannelMention } from '@model/contracts/shared/functions.js'
+import { swapMentionIDForDisplayname } from '@model/contracts/shared/functions.js'
 import {
   MESSAGE_TYPES,
-  MESSAGE_VARIANTS,
-  CHATROOM_PRIVACY_LEVEL,
-  CHATROOM_MEMBER_MENTION_SPECIAL_CHAR,
-  CHATROOM_CHANNEL_MENTION_SPECIAL_CHAR
+  MESSAGE_VARIANTS
 } from '@model/contracts/shared/constants.js'
 import { OPEN_TOUCH_LINK_HELPER } from '@utils/events.js'
-import { renderMarkdown } from '@view-utils/markdown-utils.js'
-import { htmlStringToDomObjectTree } from './chat-mentions/chat-mentions-utils.js'
 
-const TextObjectType = {
-  Text: 'TEXT',
-  MemberMention: 'MEMBER_MENTION',
-  ChannelMention: 'CHANNEL_MENTION'
-}
 export default ({
   name: 'MessageBase',
   mixins: [emoticonsMixins],
@@ -151,7 +109,8 @@ export default ({
     MessageActions,
     MessageReactions,
     SendArea,
-    ChatAttachmentPreview
+    ChatAttachmentPreview,
+    RenderMessageText
   },
   data () {
     return {
@@ -191,25 +150,8 @@ export default ({
     shouldRenderMarkdown: Boolean
   },
   computed: {
-    ...mapGetters([
-      'ourContactProfilesById',
-      'usernameFromID',
-      'chatRoomsInDetail'
-    ]),
-    textObjects () {
-      return this.generateTextObjectsFromText(this.text)
-    },
-    replyMessageObjects () {
-      return this.generateTextObjectsFromText(this.replyingMessage)
-    },
     hasAttachments () {
       return Boolean(this.attachments?.length)
-    },
-    possibleMentions () {
-      return [
-        ...Object.keys(this.ourContactProfilesById).map(u => makeMentionFromUserID(u).me).filter(v => !!v),
-        ...Object.values(this.chatRoomsInDetail).map(details => makeChannelMention(details.id))
-      ]
     }
   },
   methods: {
@@ -250,92 +192,6 @@ export default ({
     },
     openMenu () {
       this.$refs.messageAction.$refs.menu.handleTrigger()
-    },
-    isText (o) {
-      return o.type === TextObjectType.Text
-    },
-    isMemberMention (o) {
-      return o.type === TextObjectType.MemberMention
-    },
-    isChannelMention (o) {
-      return o.type === TextObjectType.ChannelMention
-    },
-    generateTextObjectsFromText (text) {
-      const containsMentionChar = str => new RegExp(`[${CHATROOM_MEMBER_MENTION_SPECIAL_CHAR}${CHATROOM_CHANNEL_MENTION_SPECIAL_CHAR}]`, 'g').test(text)
-
-      if (!text) {
-        return []
-      } else if (!containsMentionChar(text)) {
-        console.log('!@# raw html text before conversion: ', renderMarkdown(text))
-        console.log('!@# converted text: ', htmlStringToDomObjectTree(renderMarkdown(text)))
-        return [
-          {
-            type: TextObjectType.Text,
-            text: this.shouldRenderMarkdown ? renderMarkdown(text) : text
-          }
-        ]
-      }
-      const allMention = makeMentionFromUserID('').all
-
-      if (this.shouldRenderMarkdown) {
-        text = renderMarkdown(text)
-        console.log('!@# raw html text before conversion: ', text)
-        console.log('!@# converted text: ', htmlStringToDomObjectTree(text))
-      }
-
-      return text
-        // We try to find all the mentions and render them as mentions instead
-        // of regular text. The `(?<=\\s|^)` part ensures that a mention is
-        // preceded by a space or is the start of a line and the `(?=[^\\w\\d]|$)`
-        // ensures that it's followed by an end-of-line or a character that's not
-        // a letter or a number (so `Hi @user!` works).
-        .split(new RegExp(`(?<=\\s|^)(${allMention}|${this.possibleMentions.join('|')})(?=[^\\w\\d]|$)`))
-        .map(t => {
-          const genDefaultTextObj = (text) => ({
-            type: TextObjectType.Text, text
-          })
-          const genChannelMentionObj = (text) => {
-            const chatRoomID = text.slice(1)
-            const found = Object.values(this.chatRoomsInDetail).find(details => details.id === chatRoomID)
-
-            return found
-              ? {
-                  type: TextObjectType.ChannelMention,
-                  text: found.name,
-                  icon: found.privacyLevel === CHATROOM_PRIVACY_LEVEL.PRIVATE ? 'lock' : 'hashtag',
-                  disabled: found.privacyLevel === CHATROOM_PRIVACY_LEVEL.PRIVATE && !found.joined,
-                  chatRoomID: found.id
-                }
-              : genDefaultTextObj(text)
-          }
-
-          const genMemberMentionObj = (text) => {
-            const userID = text.slice(1)
-            return {
-              type: TextObjectType.MemberMention,
-              text: CHATROOM_MEMBER_MENTION_SPECIAL_CHAR + this.usernameFromID(userID),
-              userID,
-              toMe: userID === this.currentUserID
-            }
-          }
-
-          if (t === allMention) {
-            return { type: TextObjectType.MemberMention, text: t, toMe: true }
-          }
-
-          return this.possibleMentions.includes(t)
-            ? t.startsWith(CHATROOM_MEMBER_MENTION_SPECIAL_CHAR)
-              ? genMemberMentionObj(t)
-              : genChannelMentionObj(t)
-            : genDefaultTextObj(t)
-        })
-    },
-    navigateToChatRoom (obj) {
-      if (obj.disabled || obj.chatRoomID === this.$route.params?.chatRoomID) { return }
-      this.$router.push({
-        name: 'GroupChatConversation',
-        params: { chatRoomID: obj.chatRoomID }
-      })
     },
     longPressHandler (e) {
       const targetEl = e.target
@@ -471,79 +327,5 @@ export default ({
 
 .c-disappeared {
   animation: disappeared 0.5s linear;
-}
-
-.c-replying {
-  border-left: 2px;
-  border-color: #dbdbdb; // var(--text_1);
-  border-style: none none none solid;
-  font-size: 0.75rem;
-  color: var(--text_1);
-  font-style: italic;
-  padding-left: 0.25rem;
-
-  &:hover {
-    cursor: pointer;
-    color: var(--text_2);
-    border-color: var(--text_1); // var(--text_2);
-  }
-
-  .c-member-mention,
-  .c-channel-mention {
-    background-color: transparent;
-  }
-}
-
-.c-edited {
-  margin-left: 0.2rem;
-  font-size: 0.7rem;
-  color: var(--text_1);
-}
-
-.c-mention-profile-card-wrapper {
-  display: inline-block;
-}
-
-.c-member-mention,
-.c-channel-mention {
-  background-color: $primary_2;
-  color: $primary_0;
-  padding: 0 0.1rem 0.1rem;
-}
-
-.c-member-mention.c-mention-to-me {
-  background-color: $warning_1;
-}
-
-.c-channel-mention {
-  cursor: pointer;
-  transition: color 150ms;
-  outline: none;
-
-  &:hover,
-  &:focus {
-    text-decoration: underline;
-  }
-
-  &:focus {
-    color: $text_1;
-  }
-
-  &.is-disabled {
-    cursor: inherit;
-    background-color: $general_1;
-    color: $text_1;
-
-    &:hover,
-    &:focus {
-      text-decoration: none;
-      background-color: $general_1;
-    }
-  }
-
-  i {
-    font-size: 0.75em;
-    margin-right: 2px;
-  }
 }
 </style>
